requirements.txt
asgiref==3.5.2
Django==4.0.5
django-shortuuidfield==0.1.3
djangorestframework==3.13.1
pytz==2022.1
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
